+++ frontend/components/panels/WithdrawPanel.tsx
@@ -1,4 +1,3 @@
-import { tabDescriptions } from 'pages/staking'
 import { BasePanel } from './BasePanel'
 import { StakeAccount } from '@pythnetwork/staking'
 import { useBalance } from 'hooks/useBalance'
@@ -8,6 +7,10 @@ import { useStakeConnection } from 'hooks/useStakeConnection'
 type WithdrawPanelProps = {
   mainStakeAccount: StakeAccount | undefined
 }
+
+const Description =
+  'Withdraw PYTH. Transfer withdrawable tokens from the program to your wallet.'
+
 export function WithdrawPanel({ mainStakeAccount }: WithdrawPanelProps) {
   // call deposit and lock api when deposit button is clicked (create stake account if not already created)
   const withdrawMutation = useWithdrawMutation()
@@ -18,7 +21,7 @@ export function WithdrawPanel({ mainStakeAccount }: WithdrawPanelProps) {
 
   return (
     <BasePanel
-      description={tabDescriptions.Withdraw}
+      description={Description}
       tokensLabel={'Balance'}
       onAction={(amount) =>
         withdrawMutation.mutate({
@@ -28,7 +31,7 @@ export function WithdrawPanel({ mainStakeAccount }: WithdrawPanelProps) {
           stakeConnection: stakeConnection!,
         })
       }
-      actionLabel={'Locked Tokens'}
+      actionLabel={'Withdraw'}
       isActionLoading={withdrawMutation.isLoading}
       isBalanceLoading={isLoading}
       balance={unlockedPythBalance}
